pull test explorer out into its own file
