+++ npm/qsharp/src/compiler/compiler.ts
@@ -127,7 +127,7 @@ export class Compiler implements ICompiler {
       (uri: string, version: number | undefined, errors: VSDiagnostic[]) => {
         diags = errors;
       },
-      (callables: [string, ILocation][]) => { 
+      (callables: [string, ILocation][]) => {
         // do nothing; test callables are not reported in checkCode
       },
       {
